// src/features/ReservationsRooms/model/types/type.ts
export interface IReservationsData {
  //costing form
  numAdults: number
  numTeenagers: number
  numChildren: number
  roomType: string
  numNights: number
  insurance: boolean
  total: number
  // customer data
  surname: string
  name: string
  patronymic: string
  phoneNumber: string
  birthDate: string
}